+++ JazInterpreter/JazInterpreter/RelationalOperators.cs
@@ -2,72 +2,72 @@
 
 namespace JazInterpreter
 {
-	public class RelationalOperators : IRelationalOperators
-	{
-		private readonly IStackManipulation StackManipulation;
+    public class RelationalOperators : IRelationalOperators
+    {
+        private readonly IStackManipulation StackManipulation;
 
-		public RelationalOperators (IStackManipulation stackManipulation)
-		{
-			this.StackManipulation = stackManipulation;
-		}
+        public RelationalOperators(IStackManipulation stackManipulation)
+        {
+            this.StackManipulation = stackManipulation;
+        }
 
-		private Tuple<int, int> GetTopTwoValues()
-		{
-			int firstValue = PeekAndPop ();
-			int secondValue = PeekAndPop ();
+        private Tuple<int, int> GetTopTwoValues()
+        {
+            int firstValue = PeekAndPop();
+            int secondValue = PeekAndPop();
 
-			return Tuple.Create (secondValue, firstValue);
-		}
+            return Tuple.Create(secondValue, firstValue);
+        }
 
-		private int PeekAndPop()
-		{
-			int value = StackManipulation.Peek ();
-			StackManipulation.Pop ();
+        private int PeekAndPop()
+        {
+            int value = (int)StackManipulation.Peek();
+            StackManipulation.Pop();
 
-			return value;
-		}
+            return value;
+        }
 
-		public bool equal()
-		{
-			var values = GetTopTwoValues ();
+        public bool equal()
+        {
+            var values = GetTopTwoValues();
 
-			return values.Item1 == values.Item2;
-		}
+            return values.Item1 == values.Item2;
+        }
 
-		public bool lessThanOrEqualTo()
-		{
-			var values = GetTopTwoValues ();
+        public bool lessThanOrEqualTo()
+        {
+            var values = GetTopTwoValues();
 
-			return values.Item1 <= values.Item2;
-		}
+            return values.Item1 <= values.Item2;
+        }
 
-		public bool greaterThanOrEqualTo()
-		{
-			var values = GetTopTwoValues ();
+        public bool greaterThanOrEqualTo()
+        {
+            var values = GetTopTwoValues();
 
-			return values.Item1 >= values.Item2;
-		}
+            return values.Item1 >= values.Item2;
+        }
 
-		public bool lessThan()
-		{
-			var values = GetTopTwoValues ();
+        public bool lessThan()
+        {
+            var values = GetTopTwoValues();
 
-			return values.Item1 < values.Item2;
-		}
+            return values.Item1 < values.Item2;
+        }
 
-		public bool greaterThan()
-		{
-			var values = GetTopTwoValues ();
+        public bool greaterThan()
+        {
+            var values = GetTopTwoValues();
 
-			return values.Item1 > values.Item2;
-		}
+            return values.Item1 > values.Item2;
+        }
 
-		public bool otherEqual()
-		{
-			var values = GetTopTwoValues ();
+        public bool otherEqual()
+        {
+            var values = GetTopTwoValues();
 
-			return values.Item1 == values.Item2;
-		}
-	}
+            return values.Item1 == values.Item2;
+        }
+    }
 }
